feat: allow user to change themes
