+++ app/src/lib/stores/settings.ts
@@ -1,3 +1,5 @@
+import { parseTheme } from '$lib/themes';
+import { defaultDarkThemeString } from '$lib/themes/default';
 import { writable } from 'svelte/store';
 
 const defaultSettings = {
@@ -7,3 +9,6 @@ const defaultSettings = {
 };
 
 export const settings = writable(defaultSettings);
+
+export const theme = writable(parseTheme(defaultDarkThemeString));
+export const themeString = writable(defaultDarkThemeString);
